+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [connect] Add method `add_to_queue` to `Spirc` to add a track to the queue
+- [connect] Add method `add_to_queue` to `Spirc` to add tracks, episodes, albums, playlists, artists, and shows to the queue
 
 ### Changed